% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{kt_characters}
\alias{kt_characters}
\title{List characters}
\usage{
kt_characters(faction = NULL)
}
\arguments{
\item{faction}{Character. Filter by faction. Default \code{NULL} returns all.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} of character reference data.
}
\description{
List characters
}
\examples{
\dontrun{
kt_characters()
kt_characters(faction = "Autobot")
}
}
