% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{kt_connection}
\alias{kt_connection}
\title{Get DuckDB connection to kayoutf data}
\usage{
kt_connection()
}
\value{
A \code{\link[DBI:dbConnect]{DBI::dbConnect()}} DuckDB connection object.
}
\description{
Returns a singleton DuckDB connection with all Parquet tables registered
as views. The connection is created on first use and reused for subsequent
calls. It is automatically closed when the package is unloaded.
}
\examples{
\dontrun{
con <- kt_connection()
DBI::dbListTables(con)
DBI::dbGetQuery(con, "SELECT * FROM cards WHERE set_code = 'TFEU01' LIMIT 5")
}
}
