% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{kt_rarities}
\alias{kt_rarities}
\title{List rarity tiers}
\usage{
kt_rarities(set = NULL)
}
\arguments{
\item{set}{Character. Filter by set code. Default \code{NULL} returns all.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} of rarity definitions.
}
\description{
List rarity tiers
}
\examples{
\dontrun{
kt_rarities()
kt_rarities(set = "TFEU01")
}
}
