% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access.R
\name{kt_sets}
\alias{kt_sets}
\title{List all card sets}
\usage{
kt_sets()
}
\value{
A \link[tibble:tibble]{tibble::tibble} of set metadata.
}
\description{
List all card sets
}
\examples{
\dontrun{
kt_sets()
}
}
