% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_card_id}
\alias{make_card_id}
\title{Generate a card ID}
\usage{
make_card_id(set_code, rarity_code, number)
}
\arguments{
\item{set_code}{Set code (e.g. "TFEU01")}

\item{rarity_code}{Rarity code (e.g. "SSR")}

\item{number}{Card number (integer or character)}
}
\value{
Character string card ID (e.g. "TFEU01-SSR-001")
}
\description{
Generate a card ID
}
\keyword{internal}
