% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_product_id}
\alias{make_product_id}
\title{Generate a product ID}
\usage{
make_product_id(set_code, product_suffix)
}
\arguments{
\item{set_code}{Set code}

\item{product_suffix}{Product suffix (e.g. "super", "elite")}
}
\value{
Character string product ID
}
\description{
Generate a product ID
}
\keyword{internal}
