% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_rarity_id}
\alias{make_rarity_id}
\title{Generate a rarity ID}
\usage{
make_rarity_id(set_code, rarity_code)
}
\arguments{
\item{set_code}{Set code}

\item{rarity_code}{Rarity code}
}
\value{
Character string rarity ID
}
\description{
Generate a rarity ID
}
\keyword{internal}
