% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{register_parquet_tables}
\alias{register_parquet_tables}
\title{Register Parquet files as DuckDB views}
\usage{
register_parquet_tables(con)
}
\arguments{
\item{con}{A DuckDB connection}
}
\description{
Register Parquet files as DuckDB views
}
\keyword{internal}
